+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
+## [Unreleased]
+* Create a job to delete the expired session nonce's
+
 ## [0.5.0] First public release